refactor: rename User to UserView
