+++ client/components/index.js
@@ -14,4 +14,4 @@ export {default as temp} from './temp'
 export {default as NewItemForm} from './NewItemForm'
 export {default as SingleView} from './singleView'
 export {default as AllUsers} from './AllUsers'
-export {default as User} from './User'
+export {default as UserView} from './UserView'
